src/agent-reasoning/requirements.txt
langchain==0.3.27
langchain-core==0.3.74
langchain-groq==0.3.7
python-dotenv==1.1.1